+++ stytra/stimulation/stimuli/closed_loop.py
@@ -13,6 +13,7 @@ from stytra.stimulation.stimuli import (
     BackgroundStimulus,
     PositionStimulus,
     InterpolatedStimulus,
+    RadialSineStimulus
 )
 
 
@@ -392,8 +393,10 @@ class ConditionalWrapper(DynamicStimulus):
 
 
 class CenteringWrapper(ConditionalWrapper):
-    def __init__(self, *args, margin=200, **kwargs):
-        super().__init__(*args, **kwargs)
+    def __init__(self, stimulus, *args, centering_stimulus=None, margin=400, **kwargs):
+        super().__init__(*args, stim_false=stimulus,
+                         stim_true=centering_stimulus or RadialSineStimulus(stimulus.duration),
+                         **kwargs)
         self.name = "centering"
         self.margin = margin ** 2
         self.xc = 320